mdbx: fix added limit's API.

+++ mdbx.h
@@ -1691,8 +1691,8 @@ LIBMDBX_API int mdbx_is_dirty(const MDBX_txn *txn, const void *ptr);
 LIBMDBX_API int mdbx_dbi_sequence(MDBX_txn *txn, MDBX_dbi dbi, uint64_t *result,
                                   uint64_t increment);
 
-LIBMDBX_API intptr_t mdbx_limits_pgsize_min(void);
-LIBMDBX_API intptr_t mdbx_limits_pgsize_max(void);
+LIBMDBX_API int mdbx_limits_pgsize_min(void);
+LIBMDBX_API int mdbx_limits_pgsize_max(void);
 LIBMDBX_API intptr_t mdbx_limits_dbsize_min(intptr_t pagesize);
 LIBMDBX_API intptr_t mdbx_limits_dbsize_max(intptr_t pagesize);
 LIBMDBX_API intptr_t mdbx_limits_keysize_max(intptr_t pagesize);
